// typings/api.d.ts
export type InternalPlayerRatings = {
  Name: string;
  Abbr: string;
  PlayerID: number;
  LeagueID: number;
  SeasonID: number;
  G: number;
  LD: number;
  RD: number;
  LW: number;
  C: number;
  RW: number;
  Aggression: number;
  Bravery: number;
  Determination: number;
  Teamplayer: number;
  Leadership: number;
  Temperament: number;
  Professionalism: number;
  MentalToughness: number;
  GoalieStamina: number;
  Acceleration: number;
  Agility: number;
  Balance: number;
  Speed: number;
  Stamina: number;
  Strength: number;
  Fighting: number;
  Screening: number;
  GettingOpen: number;
  Passing: number;
  PuckHandling: number;
  ShootingAccuracy: number;
  ShootingRange: number;
  OffensiveRead: number;
  Checking: number;
  Faceoffs: number;
  Hitting: number;
  Positioning: number;
  ShotBlocking: number;
  Stickchecking: number;
  DefensiveRead: number;
  GPositioning: number;
  GPassing: number;
  GPokecheck: number;
  Blocker: number;
  Glove: number;
  Rebound: number;
  Recovery: number;
  GPuckhandling: number;
  LowShots: number;
  GSkating: number;
  Reflexes: number;
  Skating: number;
  Shooting: number;
  Playmaking: number;
  Defending: number;
  Physicality: number;
  Conditioning: number;
  Character: number;
  HockeySense: number;
  GoalieTechnique: number;
  GoalieOverallPositioning: number;
};